Abstracted out honeypot field, applied to both forms, cleaned up querystring
